requirements.txt
langchain
python-dotenv
langchain_community
ipykernel
bs4
arxiv
pymupdf
wikipedia
jupyter
pypdf
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit